enable ood notebook

+++ chslauncher.sh
@@ -6,20 +6,26 @@ root_dir=$(dirname "$(realpath "$0")")
 IFS='/' read -ra root_dir_list <<< "$root_dir"
 data_dir="/${root_dir_list[1]}/${root_dir_list[2]}"
 
-echo "Name Your Singularity Folder: "
+used_for_ood=0
+install_new_env=1
+
+GREEN='\033[0;32m'
+NC='\033[0m' # No Color
+
+echo "Singularity Folder Name: (type anything if you want to create a new one, no space allowed)"
 read folder_name
 
 env_dir="$root_dir/$folder_name"
 
 # Check if folder exists
 if [ -d $env_dir ]; then
-    echo "$folder_name is found. Select the next operation":
-    options=("start singularity env in read only mode" "start singularity env in read and write mode" "reinstall singularity env")
+    echo "$folder_name is found."
+    singularity_file=$(find "$env_dir" -type f -name "*cuda*" | head -n 1)
+    overlay=$(find "$env_dir" -type f -name "*overlay*" | head -n 1)
+    options=("start singularity env in read only mode" "start singularity env in read and write mode" "setup this environment for jupyter notebook in OOD" "reinstall singularity env")
     select opt in "${options[@]}"; do
         case $opt in
             "start singularity env in read only mode")
-                singularity_file=$(find "$env_dir" -type f -name "*cuda*" | head -n 1)
-                overlay=$(find "$env_dir" -type f -name "*overlay*" | head -n 1)
                 echo "Entering singularity $folder_name in read only mode"
                 echo "Always remember to activate your conda environment by typing: source /ext3/env.sh"
                 singularity exec --nv --bind $data_dir --overlay $overlay:ro $singularity_file /bin/bash
@@ -27,28 +33,70 @@ if [ -d $env_dir ]; then
                 break
                 ;;
             "start singularity env in read and write mode")
-                singularity_file=$(find "$env_dir" -type f -name "*cuda*" | head -n 1)
-                overlay=$(find "$env_dir" -type f -name "*overlay*" | head -n 1)
                 echo "Entering singularity $folder_name in read and write mode"
                 echo "Always remember to activate your conda environment by typing: source /ext3/env.sh"
                 singularity exec --nv --bind $data_dir --overlay $overlay:rw $singularity_file /bin/bash
                 exit 1
                 break
                 ;;
+            "setup this environment for jupyter notebook in OOD")
+                install_new_env=0
+                used_for_ood=1
+                break
+                ;;
             "reinstall singularity env")
-                warnings=("Yes, delete the current folder" "No, get me out of here")
-                select warn in "${warnings[@]}"; do
-                    case $warn in
-                        "Yes, delete the current folder")
-                            rm -rf $env_dir
-                            break
-                            ;;
-                        *)
-                            exit 1
-                            break
-                            ;;
-                    esac
-                done
+                read -p "All python packages installed in this singularity environment will be removed. Are you sure to reinstall it? [y]/n: " warning
+                # Convert the input to lowercase
+                warning=$(echo "$warning" | tr '[:upper:]' '[:lower:]')
+
+                # Check if the input is 'y' or 'yes', set boolean variable accordingly
+                if [[ "$warning" == "y" || "$warning" == "yes" ]]; then
+                    rm -rf $env_dir
+                    rm -rf ~/.local/share/jupyter/kernels/$folder_name
+                else
+                    exit 1
+                fi
+                break
+                ;;
+            *) 
+                echo "Invalid option"
+                exit 1
+                ;;
+        esac
+    done
+else
+    echo "No existing $folder_name found. Start creating a new one"
+fi
+
+if [ $install_new_env -eq 1 ]; then
+    # Create folder
+    mkdir $env_dir
+
+    # Check if folder creation was successful
+    if [ $? -eq 0 ]; then
+        echo "Folder \"$folder_name\" created successfully."
+    else
+        echo "Failed to create folder \"$folder_name\"."
+        exit 1
+    fi
+
+    # choose what cuda version for your ubuntu system
+    echo "Choose the cuda version":
+    options=("cuda 11.3" "cuda 11.6" "cuda 11.8")
+    select opt in "${options[@]}"; do
+        case $opt in
+            "cuda 11.3")
+                singularity_file="cuda11.3.0-cudnn8-devel-ubuntu20.04.sif"
+                # Add actions for the "large" option here
+                break
+                ;;
+            "cuda 11.6")
+                singularity_file="cuda11.6.124-cudnn8.4.0.27-devel-ubuntu20.04.4.sif"
+                # Add actions for the "small" option here
+                break
+                ;;
+            "cuda 11.8")
+                singularity_file="cuda11.8.86-cudnn8.7-devel-ubuntu22.04.2.sif"
                 break
                 ;;
             *) 
@@ -58,105 +106,101 @@ if [ -d $env_dir ]; then
         esac
     done
 
+    cp -rp /scratch/work/public/singularity/$singularity_file $env_dir
+
+
+    # choose what singularity overlay
+    echo "Choose the size of overlay, larger overlay allows you to install more python packages:"
+    options=("overlay-25GB-500K.ext3" "overlay-50G-10M.ext3")
+    select overlay in "${options[@]}"; do
+        case $overlay in
+            "overlay-25GB-500K.ext3")
+                echo "Medium Size Overlay Selected"
+                # Add actions for the "large" option here
+                break
+                ;;
+            "overlay-50G-10M.ext3")
+                echo "Large Size Overlay Selected"
+                # Add actions for the "small" option here
+                break
+                ;;
+            *) 
+                echo "Invalid option"
+                exit 1
+                ;;
+        esac
+    done
+
+    cp -rp /scratch/work/public/overlay-fs-ext3/$overlay.gz $env_dir
+    echo "unzipping your singularity $overlay, it will take a long time, please be patient"
+    gunzip $env_dir/$overlay.gz
+    echo "unzip finished"
+
+
+    # start overlay and download conda
+    overlay=$env_dir/$overlay
+    singularity_file=$env_dir/$singularity_file
+    singularity exec --nv --bind $data_dir --overlay $overlay:rw $singularity_file /bin/bash -c "
+    # download and install miniconda
+    wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
+    bash Miniconda3-latest-Linux-x86_64.sh -b -p /ext3/miniconda3
+    rm Miniconda3-latest-Linux-x86_64.sh
+
+    wget -O /ext3/env.sh https://raw.githubusercontent.com/RicercarG/NYU-Greene-HPC-Cheatsheet/main/env.sh
+
+    # init conda
+    source /ext3/env.sh
+
+    echo 'setting up base conda environment'
+
+    conda update -n base conda -y
+    conda clean --all --yes
+    conda install pip -y
+    conda install ipykernel -y
+
+    unset -f which
+
+    echo 'conda installed'
+    which conda
+    which pip
+    "
+
+    read -p "Do you want to use this python environment in open on demand jupyter notebook? [y]/n: " answer
+    # Convert the input to lowercase
+    answer=$(echo "$answer" | tr '[:upper:]' '[:lower:]')
+
+    # Check if the input is 'y' or 'yes', set boolean variable accordingly
+    if [[ "$answer" == "y" || "$answer" == "yes" ]]; then
+        used_for_ood=1
+    else
+        used_for_ood=0
+    fi
 fi
 
-# Create folder
-echo "No existing $folder_name found. Start creating a new one"
-mkdir $env_dir
+if [ $used_for_ood -eq 1 ]; then
+    mkdir -p ~/.local/share/jupyter/kernels
+    cd ~/.local/share/jupyter/kernels
+    cp -R /share/apps/mypy/src/kernel_template ./$folder_name # this should be the name of your Singularity env
+    cd ./$folder_name
+    old_word="/path/to/overlay.ext3"
+    new_word="$overlay"
+    sed -i "s#$old_word#$new_word#g" "./python"
+    old_word="/scratch/work/public/singularity/OS.sif"
+    new_word="$singularity_file"
+    sed -i "s#$old_word#$new_word#g" "./python"
 
-# Check if folder creation was successful
-if [ $? -eq 0 ]; then
-    echo "Folder \"$folder_name\" created successfully."
-else
-    echo "Failed to create folder \"$folder_name\"."
-    exit 1
+    old_word="PYTHON_LOCATION"
+    new_word="/home/$USER/.local/share/jupyter/kernels/$folder_name/python"
+    sed -i "s#$old_word#$new_word#g" "./kernel.json"
+    old_word="KERNEL_DISPLAY_NAME"
+    new_word="$folder_name"
+    sed -i "s#$old_word#$new_word#g" "./kernel.json"
 fi
 
-# choose what cuda version for your ubuntu system
-echo "Choose the cuda version":
-options=("cuda 11.3" "cuda 11.6" "cuda 11.8")
-select opt in "${options[@]}"; do
-    case $opt in
-        "cuda 11.3")
-            singularity_file="cuda11.3.0-cudnn8-devel-ubuntu20.04.sif"
-            # Add actions for the "large" option here
-            break
-            ;;
-        "cuda 11.6")
-            singularity_file="cuda11.6.124-cudnn8.4.0.27-devel-ubuntu20.04.4.sif"
-            # Add actions for the "small" option here
-            break
-            ;;
-        "cuda 11.8")
-            singularity_file="cuda11.8.86-cudnn8.7-devel-ubuntu22.04.2.sif"
-            break
-            ;;
-        *) 
-            echo "Invalid option"
-            exit 1
-            ;;
-    esac
-done
-
-cp -rp /scratch/work/public/singularity/$singularity_file $env_dir
+echo -e "${GREEN}You are all set!${NC}"
+echo -e "To start the singularity, type ${GREEN}./chslauncher.sh${NC} to run this script again"
+echo "You can also manually start this singularity using:"
+echo "singularity exec --nv --bind $data_dir --overlay $overlay.ext3:rw $singularity_file /bin/bash"
 
 
-# choose what singularity overlay
-echo "Choose the size of overlay, larger overlay allows you to install more python packages:"
-options=("overlay-25GB-500K" "overlay-50G-10M")
-select overlay in "${options[@]}"; do
-    case $overlay in
-        "overlay-25GB-500K")
-            echo "Medium Size Overlay Selected"
-            # Add actions for the "large" option here
-            break
-            ;;
-        "overlay-50G-10M")
-            echo "Large Size Overlay Selected"
-            # Add actions for the "small" option here
-            break
-            ;;
-        *) 
-            echo "Invalid option"
-            exit 1
-            ;;
-    esac
-done
 
-cp -rp /scratch/work/public/overlay-fs-ext3/$overlay.ext3.gz $env_dir
-echo "unzipping your singularity $overlay, it will take a long time, please be patient"e
-gunzip $env_dir/$overlay.ext3.gz
-echo "unzip finished"
-
-
-# start overlay and download conda
-singularity exec --nv --bind $data_dir --overlay $env_dir/$overlay.ext3:rw $env_dir/$singularity_file /bin/bash -c "
-
-# download and install miniconda
-wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-bash Miniconda3-latest-Linux-x86_64.sh -b -p /ext3/miniconda3
-rm Miniconda3-latest-Linux-x86_64.sh
-
-wget -O /ext3/env.sh https://raw.githubusercontent.com/RicercarG/NYU-Greene-HPC-Cheatsheet/main/env.sh
-
-# init conda
-source /ext3/env.sh
-
-echo 'setting up base conda environment'
-
-conda update -n base conda -y
-conda clean --all --yes
-conda install pip -y
-conda install ipykernel -y
-
-unset -f which
-
-echo 'conda installed'
-which conda
-which pip
-"
-
-echo "You are all set! To start the singularity, type the following:"
-echo "singularity exec --nv --bind $data_dir --overlay $env_dir/$overlay.ext3:rw $env_dir/$singularity_file /bin/bash"
-echo "To quit the singularity, simply type exit in command line"
-echo "You can start always this singularity environment by rerunning this file and type in the environment name"
